Refactor help plugin
