+++ templates/help_response.j2
@@ -1,1 +1,5 @@
-Help functionality coming on
+```
+SHOW HELP MENU: help
+
+<More help functionality will be updated as features are added>
+```
