+++ tests/test_help_response.py
@@ -1,0 +1,6 @@
+import unittest
+from plugins import help_plugin
+
+
+class TestHelpFunction(unittest.TestCase):
+    pass
